+++ core-services/service-request/src/main/java/org/egov/servicerequest/validators/ServiceDefinitionRequestValidator.java
@@ -1,11 +1,8 @@
 package org.egov.servicerequest.validators;
 
 import org.egov.servicerequest.repository.ServiceDefinitionRequestRepository;
-import org.egov.servicerequest.web.models.AttributeDefinition;
-import org.egov.servicerequest.web.models.ServiceDefinition;
-import org.egov.servicerequest.web.models.ServiceDefinitionCriteria;
-import org.egov.servicerequest.web.models.ServiceDefinitionRequest;
-import org.egov.servicerequest.web.models.ServiceDefinitionSearchRequest;
+import org.egov.servicerequest.repository.ServiceRequestRepository;
+import org.egov.servicerequest.web.models.*;
 import org.egov.tracer.model.CustomException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
@@ -18,20 +15,16 @@ import java.util.List;
 import java.util.Set;
 import java.util.regex.Pattern;
 
-import static org.egov.servicerequest.error.ErrorCode.ATTRIBUTE_CODE_UNIQUENESS_ERR_CODE;
-import static org.egov.servicerequest.error.ErrorCode.ATTRIBUTE_CODE_UNIQUENESS_ERR_MSG;
-import static org.egov.servicerequest.error.ErrorCode.INVALID_ATTRIBUTE_DEFINITION_ERR_CODE;
-import static org.egov.servicerequest.error.ErrorCode.INVALID_ATTRIBUTE_DEFINITION_ERR_MSG;
-import static org.egov.servicerequest.error.ErrorCode.INVALID_REGEX_ERR_CODE;
-import static org.egov.servicerequest.error.ErrorCode.INVALID_REGEX_ERR_MSG;
-import static org.egov.servicerequest.error.ErrorCode.SERVICE_DEFINITION_ALREADY_EXISTS_ERR_CODE;
-import static org.egov.servicerequest.error.ErrorCode.SERVICE_DEFINITION_ALREADY_EXISTS_ERR_MSG;
+import static org.egov.servicerequest.error.ErrorCode.*;
 
 @Component
 public class ServiceDefinitionRequestValidator {
 
     @Autowired
     private ServiceDefinitionRequestRepository serviceDefinitionRequestRepository;
+
+    @Autowired
+    private ServiceRequestRepository serviceRequestRepository;
 
     public void validateServiceDefinitionRequest(ServiceDefinitionRequest serviceDefinitionRequest){
         ServiceDefinition serviceDefinition = serviceDefinitionRequest.getServiceDefinition();
@@ -93,10 +86,52 @@ public class ServiceDefinitionRequestValidator {
         }
     }
 
-    public void validateUpdateRequest(ServiceDefinitionRequest serviceDefinitionRequest) {
+    private List<ServiceDefinition> validateExistence(ServiceDefinition serviceDefinition){
+        List<ServiceDefinition> serviceDefinitionList = serviceDefinitionRequestRepository.
+          getServiceDefinitions(ServiceDefinitionSearchRequest.builder().serviceDefinitionCriteria(ServiceDefinitionCriteria.builder().tenantId(serviceDefinition.getTenantId()).code(Collections.singletonList(serviceDefinition.getCode())).build()).build());
 
-        // TO DO
+        //check if valid service definition exists
+        if(CollectionUtils.isEmpty(serviceDefinitionList)){
+            throw new CustomException(SERVICE_DEFINITION_NOT_EXIST_ERR_CODE, SERVICE_DEFINITION_NOT_EXIST_ERR_MSG);
+        }
 
+        //check if service definition is active or not
+        if(!serviceDefinitionList.get(0).getIsActive()){
+            throw new CustomException(INACTIVE_SERVICE_DEFINITION_ERR_CODE, INACTIVE_SERVICE_DEFINITION_ERR_MSG);
+        }
+
+        return serviceDefinitionList;
+    }
+
+    private void validateService(List<ServiceDefinition> serviceDefinition){
+        List<Service> service = serviceRequestRepository.getService(
+          ServiceSearchRequest.builder().serviceCriteria(
+            ServiceCriteria.builder().serviceDefIds(Collections.singletonList(serviceDefinition.get(0).getId())).build()
+          ).build()
+        );
+
+        //already a service exists corresponding to service definition
+        if(!CollectionUtils.isEmpty(service)){
+            throw new CustomException(VALID_SERVICE_EXIST_ERR_CODE, VALID_SERVICE_EXIST_ERR_MSG);
+        }
+    }
+
+    public List<ServiceDefinition> validateUpdateRequest(ServiceDefinitionRequest serviceDefinitionRequest) {
+        ServiceDefinition serviceDefinition = serviceDefinitionRequest.getServiceDefinition();
+
+        //Check if valid Service Definition exists
+        List<ServiceDefinition> serviceDefinitionList = validateExistence(serviceDefinition);
+
+        //Check if a service exists corresponding to this service definition
+        validateService(serviceDefinitionList);
+
+        validateAttributeDefinitionUniqueness(serviceDefinition);
+
+        validateAttributeValuesAsPerDataType(serviceDefinition);
+
+        validateRegex(serviceDefinition);
+
+        return serviceDefinitionList;
     }
 
 }