+++ health-services/plan-service/src/main/java/digit/web/models/PlanConfigurationSearchCriteria.java
@@ -1,6 +1,5 @@
 package digit.web.models;
 
-import jakarta.validation.Valid;
 import jakarta.validation.constraints.Max;
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotNull;
@@ -11,6 +10,8 @@ import lombok.AllArgsConstructor;
 import lombok.NoArgsConstructor;
 import lombok.Data;
 import lombok.Builder;
+
+import java.util.List;
 
 /**
  * PlanConfigurationSearchCriteria
@@ -30,14 +31,17 @@ public class PlanConfigurationSearchCriteria {
     @JsonProperty("id")
     private String id = null;
 
+    @JsonProperty("ids")
+    private List<String> ids = null;
+
     @JsonProperty("name")
     private String name = null;
 
-    @JsonProperty("executionPlanId")
-    private String executionPlanId = null;
+    @JsonProperty("campaignId")
+    private String campaignId = null;
 
     @JsonProperty("status")
-    private String status = null;
+    private List<String> status = null;
 
     @JsonProperty("userUuid")
     private String userUuid = null;
@@ -50,5 +54,4 @@ public class PlanConfigurationSearchCriteria {
     @Min(1)
     @Max(50)
     private Integer limit;
-
 }