+++ health-services/project-factory/src/server/controllers/localisationController/localisation.controller.ts
@@ -3,6 +3,7 @@ import { logger } from "../../utils/logger";
 import { httpRequest } from "../../utils/request";
 import config from "../../config/index";
 import { convertLocalisationResponseToMap } from "../../utils/localisationUtils";
+import { defaultRequestInfo } from "../../api/coreApis";
 
 let cachedResponse = {};
 
@@ -15,7 +16,7 @@ class Localisation {
   // Hold the single instance of the class
   private static instance: Localisation;
   constructor() {
-    this.localizationHost = config.host.localizationHost
+    this.localizationHost = config.host.localizationHost;
   }
   // Public method to provide access to the single instance
   public static getInstance(): Localisation {
@@ -36,12 +37,13 @@ class Localisation {
   public getLocalisedData: any = async (
     module: string,
     locale: string,
-    tenantId: string
+    tenantId: string,
+    overrideCache: boolean
   ) => {
     logger.info(
       `Checks Localisation message is available in cache for module ${module}, locale ${locale}, tenantId ${tenantId}`
     );
-    if (!this?.cachedResponse?.[`${module}-${locale}`]) {
+    if (!this?.cachedResponse?.[`${module}-${locale}`] || overrideCache) {
       logger.info(`Not found in cache`);
       await this.fetchLocalisationMessage(module, locale, tenantId);
     }
@@ -81,6 +83,34 @@ class Localisation {
     );
     cachedResponse = { ...this.cachedResponse };
   };
+  
+  // Calls the cache burst API of localization
+  public cacheBurst = async (
+  ) => {
+    logger.info(`Calling localization cache burst api`);
+    const RequestInfo = defaultRequestInfo;
+    const requestBody = {
+      RequestInfo
+    }
+    await httpRequest(
+      this.localizationHost + config.paths.cacheBurst,
+      requestBody)
+  };
+
+
+  private checkCacheAndDeleteIfExists = (module: string, locale: "string") => {
+    logger.info(
+      `Received to checkCacheAndDeleteIfExists for module ${module}, locale ${locale}`
+    );
+    if (this.cachedResponse?.[`${module}-${locale}`]) {
+      logger.info(`cache found to for module ${module}, locale ${locale}`);
+      if (delete this.cachedResponse?.[`${module}-${locale}`]) {
+        logger.info(
+          `cache deleted for module ${module}, locale ${locale}, since new data has been created`
+        );
+      }
+    }
+  };
 
   /**
    * Create localisation entries by sending a POST request to the localization host.
@@ -98,19 +128,26 @@ class Localisation {
       // Construct request body with RequestInfo and localisation messages
       const requestBody = { RequestInfo, messages, tenantId };
       // Construct URL for localization create endpoint
-      const url =
-        this.localizationHost + config.paths.localizationCreate;
+      const url = this.localizationHost + config.paths.localizationCreate;
       // Log the start of the localisation messages creation process
-      logger.info("Creating the localisation messages");
+      logger.info(`Creating the localisation messages of count ${messages?.length}`);
       // Send HTTP POST request to create localisation messages
 
       await httpRequest(url, requestBody);
+
+      messages &&
+        messages?.length > 0 &&
+        this.checkCacheAndDeleteIfExists(
+          messages?.[0]?.module,
+          messages?.[0]?.locale
+        );
       // Log the completion of the localisation messages creation process
       logger.info("Localisation messages created successfully");
     } catch (e: any) {
       // Log and handle any errors that occur during the process
       console.log(e);
       logger.error(String(e));
+      throw new Error(e);
     }
   };
 }