+++ health-services/stock/src/main/java/org/egov/stock/util/ValidatorUtil.java
@@ -6,8 +6,8 @@ import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
 
-import digit.models.coremodels.UserSearchRequest;
 import org.egov.common.contract.request.RequestInfo;
+import org.egov.common.contract.user.UserSearchRequest;
 import org.egov.common.ds.Tuple;
 import org.egov.common.models.Error;
 import org.egov.common.models.stock.SenderReceiverType;
@@ -268,6 +268,7 @@ public class ValidatorUtil {
 
 			// Get the sender and receiver IDs from the stock object
 			String senderId = stock.getSenderId();
+			String receiverId = stock.getReceiverId();
 
 			// Get the list of facility IDs mapped to the reference ID of the stock
 			List<String> facilityIds = ProjectFacilityMappingOfIds.get(stock.getReferenceId());
@@ -303,8 +304,6 @@ public class ValidatorUtil {
 					// If facilityIds are empty, log an error for both sender and receiver
 					populateErrorForStock(stock, senderId + " and " + receiverId, errorDetailsMap);
 				}
-			} else {
-				populateErrorForStock(stock, senderId, errorDetailsMap);
 			}
 		}
 	}