+++ health-services/project/CHANGELOG.md
@@ -2,6 +2,7 @@ All notable changes to this module will be documented in this file.
 
 ## 1.1.5 - 2024-08-07
 - Added UserAction functionality with support for Location capture.
+- Added a boolean flag to enable attendance features, set to true by default.
 
 ## 1.1.4 - 2024-05-29
 - Integrated Core 2.9LTS