+++ health-services/project/CHANGELOG.md
@@ -1,5 +1,8 @@
 All notable changes to this module will be documented in this file.
 
+## 1.1.6 - 2025-01-27
+- Added isAncestorProjectId param for search projects API to support search projects with ancestor project id as well
+
 ## 1.1.5 - 2024-08-07
 - Added UserAction functionality with support for Location capture.
 
@@ -13,6 +16,7 @@ All notable changes to this module will be documented in this file.
 - Upgraded PostgresSQL Driver version to 42.7.1
 - Upgraded Flyway base image version to 10.7.1 for DB Migration
 - Upgraded Flyway-Core to 9.22.3
+- Added `ExistentEntityValidator` fixes
 
 ## 1.1.2 - 2024-02-26
 - Implemented validation for updating project start date and end date.